Fixes #33967 - Loading constants in initializers/active_record_extensions.rb
