+++ config/application.rb
@@ -72,7 +72,6 @@ end
 # load the corresponding bit of fog
 require 'fog/ovirt' if defined?(::OVIRT)
 
-require_dependency File.expand_path('../app/models/application_record.rb', __dir__)
 require_dependency File.expand_path('../lib/foreman.rb', __dir__)
 require_dependency File.expand_path('../lib/timed_cached_store.rb', __dir__)
 require_dependency File.expand_path('../lib/foreman/exception', __dir__)
@@ -369,6 +368,7 @@ module Foreman
 
     def setup_auditing
       Audit.include AuditSearch
+      Audit.include HasManyCommon
     end
   end
