--- config/initializers/uri_jail.rb
@@ -1,3 +1,0 @@
-class URI::Generic::Jail < Safemode::Jail
-  allow :host, :path, :port, :query, :scheme
-end
